% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_management.R
\name{ask_yes_no}
\alias{ask_yes_no}
\title{Ask Yes/No Question}
\usage{
ask_yes_no(prompt = "Do you want to continue? (Y/N): ")
}
\arguments{
\item{prompt}{A character string specifying the question to ask.}
}
\value{
A character string "Y" or "N" based on the user's response.
}
\description{
This function prompts the user to answer a yes/no question.
}
\keyword{internal}
