% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/car.R
\name{bi.dist.car}
\alias{bi.dist.car}
\title{Conditional Autoregressive (CAR) Distribution}
\usage{
bi.dist.car(
  loc,
  correlation,
  conditional_precision,
  adj_matrix,
  is_sparse = FALSE,
  validate_args = py_none(),
  name = "x",
  obs = py_none(),
  mask = py_none(),
  sample = FALSE,
  seed = 0,
  shape = c(),
  event = 0,
  create_obj = FALSE
)
}
\arguments{
\item{loc}{Numeric vector, matrix, or array representing the mean of the distribution.}

\item{correlation}{Numeric vector, matrix, or array representing the correlation between variables.}

\item{conditional_precision}{Numeric vector, matrix, or array representing the precision of the distribution.}

\item{adj_matrix}{Numeric vector, matrix, or array representing the adjacency matrix defining the graph.}

\item{is_sparse}{Logical indicating whether the adjacency matrix is sparse. Defaults to `FALSE`.}

\item{validate_args}{Logical indicating whether to validate arguments. Defaults to `reticulate::py_none()`.}
}
\value{
- When \code{sample=FALSE}, a BI CAR distribution object (for model building).

 - When \code{sample=TRUE}, a JAX array of samples drawn from the CAR distribution (for direct sampling).

 - When \code{create_obj=TRUE}, the raw BI distribution object (for advanced use cases).
}
\description{
The CAR distribution models a vector of variables where each variable is a linear
combination of its neighbors in a graph.

\deqn{p(x) = \prod_{i=1}^{K} \mathcal{N}(x_i | \mu_i, \Sigma_i)}

where \eqn{\mu_i} is a function of the values of the neighbors of site \eqn{i}
and \eqn{\Sigma_i} is the variance of site \eqn{i}.

The CAR distribution is a special case of the multivariate normal distribution.
It is used to model spatial data, such as temperature or precipitation.
}
\examples{
\donttest{
library(BI)
m=importBI(platform='cpu')
bi.dist.car(loc = 0, correlation = 0.5, conditional_precision = 0.1, adj_matrix = 2, sample = TRUE)
}
}
