% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_copula_beta.R
\name{bi.dist.gaussian_copula_beta}
\alias{bi.dist.gaussian_copula_beta}
\title{Gaussian Copula Beta distribution.}
\usage{
bi.dist.gaussian_copula_beta(
  concentration1,
  concentration0,
  correlation_matrix = py_none(),
  correlation_cholesky = py_none(),
  validate_args = FALSE,
  name = "x",
  obs = py_none(),
  mask = py_none(),
  sample = FALSE,
  seed = 0,
  shape = c(),
  event = 0,
  create_obj = FALSE
)
}
\arguments{
\item{concentration1}{A numeric vector or matrix representing the first shape parameter of the Beta distribution.}

\item{concentration0}{A numeric vector or matrix representing the second shape parameter of the Beta distribution.}

\item{correlation_cholesky}{A numeric vector, matrix, or array representing the Cholesky decomposition of the correlation matrix.}

\item{mask}{A logical vector. Optional boolean array to mask observations.}

\item{shape}{A numeric vector.  This is used as `sample_shape` to draw a raw JAX array of the given shape when `sample=True`.}

\item{event}{Integer indicating the number of batch dimensions to reinterpret as event dimensions (used in model building).}

\item{create_obj}{Logical. If `TRUE`, returns the raw BI distribution object instead of creating a sample site.}
}
\value{
- When \code{sample=FALSE}, a BI Gaussian Copula Beta distribution object (for model building).

 - When \code{sample=TRUE}, a JAX array of samples drawn from the Gaussian Copula Beta distribution (for direct sampling).

 - When \code{create_obj=TRUE}, the raw BI distribution object (for advanced use cases).
}
\description{
Samples from a Gaussian Copula Beta distribution.
This distribution combines a Gaussian copula with a Beta distribution.
The Gaussian copula models the dependence structure between random variables,
while the Beta distribution defines the marginal distributions of each variable.
\eqn{f(x) = \int_{-\infty}^{\infty} g(x|u) h(u) du}
}
\examples{
\donttest{
library(BI)
m=importBI(platform='cpu')
bi.dist.gaussian_copula_beta(
  concentration1 = c(2.0, 3.0),
  concentration0 = c(5.0, 3.0),
  correlation_matrix = matrix(c(1.0, 0.7, 0.7, 1.0), nrow = 2, byrow = TRUE),
  sample = TRUE)
  }
}
\seealso{
This is a wrapper of  \url{https://num.pyro.ai/en/stable/distributions.html#gaussiancopulabetadistribution}
}
