% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_truncated_distribution.R
\name{bi.dist.left_truncated_distribution}
\alias{bi.dist.left_truncated_distribution}
\title{Samples from a left-truncated distribution.}
\usage{
bi.dist.left_truncated_distribution(
  base_dist,
  low = 0,
  validate_args = py_none(),
  name = "x",
  obs = py_none(),
  mask = py_none(),
  sample = FALSE,
  seed = 0,
  shape = c(),
  event = 0,
  create_obj = FALSE
)
}
\arguments{
\item{base_dist}{The base distribution to truncate. Must be univariate and have real support.}

\item{low}{The lower truncation bound. Values less than this are excluded from the distribution.}

\item{mask}{An optional boolean vector to mask observations.}

\item{shape}{A numeric vector. When \code{sample=FALSE} (model building),
this is used with `.expand(shape)` to set the distribution's batch shape.
When \code{sample=TRUE} (direct sampling), this is used as `sample_shape` to draw a raw
JAX array of the given shape.}

\item{event}{The number of batch dimensions to reinterpret as event dimensions (used in model building).}

\item{create_obj}{A logical value. If `TRUE`, returns the raw BI distribution object instead of creating a sample site.}
}
\value{
- When \code{sample=FALSE}: A BI LeftTruncatedDistribution distribution object (for model building).

   - When \code{sample=TRUE}: A JAX array of samples drawn from the LeftTruncatedDistribution distribution (for direct sampling).

   - When \code{create_obj=TRUE}: The raw BI distribution object (for advanced use cases).
}
\description{
A left-truncated distribution is a probability distribution
obtained by restricting the support of another distribution
to values greater than a specified lower bound. This is useful
when dealing with data that is known to be greater than a certain value.


\deqn{f(x) = \begin{cases}
          \frac{f(x)}{P(X > \text{low})} & \text{if } x > \text{low} \\
          0 & \text{otherwise}
          \end{cases}}
}
\examples{
\donttest{
library(BI)
m=importBI(platform='cpu')
bi.dist.left_truncated_distribution( base_dist = bi.dist.normal(loc = 1, scale = 10 ,  create_obj = TRUE),  sample = TRUE)
}
}
\seealso{
This is a wrapper of  \url{https://num.pyro.ai/en/stable/distributions.html#lefttruncateddistribution}
}
