% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_same_family.R
\name{bi.dist.mixture_same_family}
\alias{bi.dist.mixture_same_family}
\title{A finite mixture of component distributions from the same family.}
\usage{
bi.dist.mixture_same_family(
  mixing_distribution,
  component_distribution,
  validate_args = py_none(),
  name = "x",
  obs = py_none(),
  mask = py_none(),
  sample = FALSE,
  seed = 0,
  shape = c(),
  event = 0,
  create_obj = FALSE
)
}
\arguments{
\item{mask}{A logical vector, matrix, or array to mask observations.}

\item{shape}{A numeric vector specifying the shape of the distribution.}

\item{event}{Integer representing the number of batch dimensions to reinterpret as event dimensions (used in model building).}

\item{create_obj}{Logical; If TRUE, returns the raw BI distribution object instead of creating a sample site.}

\item{loc}{A numeric vector, matrix, or array representing the location parameter of the component distribution.}

\item{scale}{A numeric vector, matrix, or array representing the scale parameter of the component distribution.}
}
\value{
- When \code{sample=FALSE}, a BI MixtureSameFamily distribution object (for model building).

   - When \code{sample=TRUE}, a JAX array of samples drawn from the MixtureSameFamily distribution (for direct sampling).

   - When \code{create_obj=TRUE}, the raw BI distribution object (for advanced use cases).
}
\description{
This mixture only supports a mixture of component distributions that are all
of the same family. The different components are specified along the last
batch dimension of the input ``component_distribution``. If you need a
mixture of distributions from different families, use the more general
implementation in :class:`bi.dist.mixture_general`.
}
\examples{
\donttest{
library(BI)
m=importBI(platform='cpu')
bi.dist.mixture_same_family(
mixing_distribution = bi.dist.categorical(probs = c(0.3, 0.7),create_obj = TRUE),
component_distribution = bi.dist.normal(0,1, shape = c(2), create_obj = TRUE),
sample = TRUE)
}
}
\seealso{
This is a wrapper of  \url{https://num.pyro.ai/en/stable/distributions.html#mixture-same-family}
}
