% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wishart_cholesky.R
\name{bi.dist.wishart_cholesky}
\alias{bi.dist.wishart_cholesky}
\title{Wishart Cholesky Distribution}
\usage{
bi.dist.wishart_cholesky(
  concentration,
  scale_matrix = py_none(),
  rate_matrix = py_none(),
  scale_tril = py_none(),
  validate_args = py_none(),
  name = "x",
  obs = py_none(),
  mask = py_none(),
  sample = FALSE,
  seed = 0,
  shape = c(),
  event = 0,
  create_obj = FALSE
)
}
\arguments{
\item{concentration}{(numeric or vector) Positive concentration parameter analogous to the
concentration of a `Gamma` distribution. The concentration must be larger
than the dimensionality of the scale matrix.}

\item{scale_matrix}{(numeric vector, matrix, or array, optional) Scale matrix analogous to the inverse rate of a `Gamma`
distribution. If not provided, `rate_matrix` or `scale_tril` must be.}

\item{rate_matrix}{(numeric vector, matrix, or array, optional) Rate matrix anaologous to the rate of a `Gamma`
distribution. If not provided, `scale_matrix` or `scale_tril` must be.}

\item{scale_tril}{(numeric vector, matrix, or array, optional) Cholesky decomposition of the `scale_matrix`.
If not provided, `scale_matrix` or `rate_matrix` must be.}
}
\value{
- When \code{sample=FALSE}, a BI Wishart Cholesky  distribution object (for model building).

 - When \code{sample=TRUE}, a JAX array of samples drawn from the Wishart Cholesky  distribution (for direct sampling).

 - When \code{create_obj=TRUE}, the raw BI distribution object (for advanced use cases).
}
\description{
The Wishart distribution is a multivariate distribution used as a prior distribution
for covariance matrices. This implementation represents the distribution in terms
of its Cholesky decomposition.

The Wishart distribution is a multivariate distribution used as a prior distribution
for covariance matrices. This implementation represents the distribution in terms
of its Cholesky decomposition.
}
\examples{
\donttest{
library(BI)
m=importBI(platform='cpu')
bi.dist.wishart_cholesky(sample = TRUE)
}
}
