% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_inflated_poisson.R
\name{bi.dist.zero_inflated_poisson}
\alias{bi.dist.zero_inflated_poisson}
\title{A Zero Inflated Poisson distribution.}
\usage{
bi.dist.zero_inflated_poisson(
  gate,
  rate = 1,
  validate_args = py_none(),
  name = "x",
  obs = py_none(),
  mask = py_none(),
  sample = FALSE,
  seed = 0,
  shape = c(),
  event = 0,
  create_obj = FALSE
)
}
\arguments{
\item{gate}{The gate parameter.}

\item{rate}{A numeric vector, matrix, or array representing the rate parameter of the underlying Poisson distribution.}

\item{mask}{An optional boolean vector, matrix, or array to mask observations.}

\item{sample}{Logical; If `TRUE`, draws samples from the distribution.}

\item{shape}{A numeric vector used to shape the distribution. When `sample=False` (model building), this is used with `.expand(shape)` to set the distribution's batch shape. When `sample=True` (direct sampling), this is used as `sample_shape` to draw a raw JAX array of the given shape.}

\item{event}{The number of batch dimensions to reinterpret as event dimensions (used in model building).}

\item{create_obj}{Logical; If `TRUE`, returns the raw BI distribution object instead of creating a sample site.}
}
\value{
- When \code{sample=FALSE}, a BI Zero Inflated Poisson distribution object (for model building).

 - When \code{sample=TRUE}, a JAX array of samples drawn from the Zero Inflated Poisson distribution (for direct sampling).

 - When \code{create_obj=TRUE}, the raw BI distribution object (for advanced use cases).
}
\description{
This distribution combines two Poisson processes: one with a rate parameter and another that generates only zeros.
The probability of observing a zero is determined by the 'gate' parameter, while the probability of observing a non-zero value is governed by the 'rate' parameter of the underlying Poisson distribution.

\deqn{P(X = k) = (1 - gate) * \frac{e^{-rate} rate^k}{k!} + gate}
}
\examples{
#' \donttest{
library(BI)
m=importBI(platform='cpu')
bi.dist.zero_inflated_poisson(gate=0.3, rate = 5, sample = TRUE)
}
}
