% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{bi.doc}
\alias{bi.doc}
\title{title Open the BI Documentation Website}
\usage{
bi.doc()
}
\value{
Invisibly returns `NULL`. The function is called for its side effect
of opening a URL.
}
\description{
This function opens the BI documentation webpage in the user's default
web browser.
}
