% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_management.R
\name{BI_check_presence}
\alias{BI_check_presence}
\title{Check if BayesInference Python package is installed}
\usage{
BI_check_presence()
}
\value{
Logical `TRUE` if the package is found, otherwise prints a message
  and returns `NULL`.
}
\description{
Internal function to check whether the Python package `BayesInference`
is installed in the default reticualte virtual environment.
}
\keyword{internal}
