% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_management.R
\name{BI_force_dependencies_version}
\alias{BI_force_dependencies_version}
\title{Force Installation of Specific Python Dependency Versions}
\usage{
BI_force_dependencies_version()
}
\value{
Invisible `NULL`. The function is called for its side effect of
  installing Python packages.
}
\description{
Internal function to install specific versions of Python packages required
for the package to function correctly. This ensures a consistent and stable
environment, avoiding potential conflicts from dependency updates.
}
\details{
This function uses `reticulate::py_install` to force the installation of
specific versions of `jax`, `jaxlib`, `numpyro`, `numpy`,
`tensorflow_probability`, `arviz`, `matplotlib`, `seaborn`, `pandas`, and
`scipy`. It is particularly useful for ensuring reproducibility of analyses.
}
\keyword{internal}
