% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_management.R
\name{BI_load}
\alias{BI_load}
\title{Load the BI module from Python}
\usage{
BI_load()
}
\value{
The Python object corresponding to `BI$bi`.
}
\description{
Load the BI module from Python
}
\details{
Internal helper function to import the `BI` Python module.
}
\keyword{internal}
