% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_management.R
\name{BI_starting_test}
\alias{BI_starting_test}
\title{Run a starting test for the BI environment}
\usage{
BI_starting_test()
}
\description{
Run a starting test for the BI environment
}
\details{
Internal function that checks whether Python and the `BayesInference`
package are available, and sets internal `.BI_env` flags accordingly.

Checks if the `reticulate` package is installed, if Python is available,
and if the `BayesInference` package exists in the target environment.
Updates `.BI_env$loaded` and `.BI_env$.bi` accordingly.
}
\keyword{internal}
