% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_management.R
\name{check_env}
\alias{check_env}
\title{Check if the default virtual environment is available}
\usage{
check_env()
}
\value{
A logical value indicating whether the "cpu" or "gpu" environment exists.
}
\description{
This function checks for the existence of the default Python virtual environments.
}
