% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_posterior.R
\name{convert_posterior}
\alias{convert_posterior}
\title{Convert Posterior Samples}
\usage{
convert_posterior(posteriors)
}
\arguments{
\item{posteriors}{Posterior samples from Python model.}
}
\value{
A list of posterior samples in R format.
}
\description{
Converts posterior samples from Python to R format.
}
\keyword{internal}
