% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load.R
\docType{data}
\name{.BI_env}
\alias{.BI_env}
\title{Package load hook}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.BI_env
}
\description{
Internal function to run when the package is loaded.
}
\keyword{internal}
