% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load.R
\name{.onAttach}
\alias{.onAttach}
\title{Package attach}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{Internal.}

\item{pkgname}{Internal.}
}
\description{
Internal function to run when the package is attached.
}
\keyword{internal}
