% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_management.R
\name{install_dependencies}
\alias{install_dependencies}
\title{Install dependencies}
\usage{
install_dependencies(envname = "BayesInference")
}
\arguments{
\item{envname}{A character string specifying the name of the virtual environment.}
}
\description{
Install dependencies
}
\details{
Internal helper function to install dependencies for the `BayesInference` package.
}
\keyword{internal}
