% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_management.R
\name{list_packages}
\alias{list_packages}
\title{List packages in a virtual environment}
\usage{
list_packages(envname = "BayesInference")
}
\arguments{
\item{envname}{A character string specifying the name of the virtual environment.}
}
\description{
List packages in a virtual environment
}
\details{
Internal helper function to list packages in a virtual environment.
}
\keyword{internal}
