% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_management.R
\name{remove_env_name}
\alias{remove_env_name}
\title{Remove a virtual environment}
\usage{
remove_env_name(env_name = "BayesInference")
}
\arguments{
\item{env_name}{A character string specifying the name of the virtual environment.}
}
\description{
Internal helper function to remove a virtual environment.
}
\keyword{internal}
