% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_management.R
\name{setup_env}
\alias{setup_env}
\title{Create a Python virtual environment}
\usage{
setup_env(env_name = "BayesInference", backend = "cpu")
}
\arguments{
\item{env_name}{A character string for the name of the virtual environment.
Defaults to "BayesInference-cpu" or "BayesInference-gpu".}

\item{backend}{A character string specifying the backend, either "cpu" or "gpu".}
}
\value{
The path to the created virtual environment.

The path to the created virtual environment.
}
\description{
This function creates a Python virtual environment using reticulate.
}
