% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_management.R
\name{update_BI}
\alias{update_BI}
\title{Update the BI module from Python}
\usage{
update_BI(envname = "BayesInference")
}
\arguments{
\item{envname}{A character string specifying the name of the virtual environment.}
}
\description{
Update the BI module from Python
}
\details{
Internal helper function to update the `BI` Python module.
}
\keyword{internal}
