% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eulermaruyama.R
\name{bi.dist.euler_maruyama}
\alias{bi.dist.euler_maruyama}
\title{EulerMaruyama distribution wrapper.}
\usage{
bi.dist.euler_maruyama(
  t,
  sde_fn,
  init_dist,
  validate_args = py_none(),
  name = "x",
  obs = py_none(),
  mask = py_none(),
  sample = FALSE,
  seed = 0,
  shape = c(),
  event = 0,
  create_obj = FALSE
)
}
\arguments{
\item{t}{<class 'inspect._empty'>}

\item{sde_fn}{<class 'inspect._empty'>}

\item{init_dist}{<class 'inspect._empty'>}

\item{validate_args}{None}

\item{mask}{(jnp.ndarray, bool): Optional boolean array to mask observations. This is passed to the `infer={'obs_mask': ...}` argument of `numpyro.sample`.}

\item{shape}{(tuple): A multi-purpose argument for shaping. - When sample=False (model building), this is used with `.expand(shape)` to set the distribution's batch shape. - When sample=True (direct sampling), this is used as `sample_shape` to draw a raw JAX array of the given shape.}

\item{event}{(int): The number of batch dimensions to reinterpret as event dimensions (used in model building).}

\item{create_obj}{(bool): If True, returns the raw NumPyro distribution object instead of creating a sample site. This is essential for building complex distributions like `MixtureSameFamily`.}
}
\description{
EulerMaruyama distribution wrapper.
}
\examples{
library(BI)
m=importBI(platform='cpu')
m.dist.euler_maruyama(t=jnp.array([0.0, 0.1, 0.2]), sde_fn=lambda x, t: (x, 1.0), init_dist=m.dist.normal(0.0, 1.0, create_obj=True)
bi.dist.euler_maruyama(t=c(0,0.1,0.2), sde_fn = lambda x, t: (x, 1.0), .bi$dist$normal(0,1,create_obj = T), sample = TRUE, event = c(1,10))
}
