% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_nomial_logits.R
\name{bi.dist.multinomial_logits}
\alias{bi.dist.multinomial_logits}
\title{Samples from a MultinomialLogits distribution.}
\usage{
bi.dist.multinomial_logits(
  logits,
  total_count = 1,
  total_count_max = py_none(),
  validate_args = py_none(),
  name = "x",
  obs = py_none(),
  mask = py_none(),
  sample = FALSE,
  seed = 0,
  shape = c(),
  event = 0,
  create_obj = FALSE
)
}
\arguments{
\item{logits}{A numeric vector, matrix, or array representing the logits for each outcome.}

\item{total_count}{A numeric vector, matrix, or array representing the total number of trials.}

\item{mask}{A logical vector, matrix, or array to mask observations.}

\item{shape}{A numeric vector specifying the shape of the distribution.  Use a vector (e.g., `c(10)`) to define the shape.}

\item{event}{Integer specifying the number of batch dimensions to reinterpret as event dimensions (used in model building).}

\item{create_obj}{Logical; If TRUE, returns the raw BI distribution object instead of creating a sample site.}
}
\value{
- When \code{sample=FALSE}, a BI MultinomialLogits distribution object (for model building).

   - When \code{sample=TRUE}, a JAX array of samples drawn from the MultinomialLogits distribution (for direct sampling).

   - When \code{create_obj=TRUE}, the raw BI distribution object (for advanced use cases).
}
\description{
This distribution represents the probability of observing a specific outcome from a multinomial experiment,
given the logits for each outcome. The logits are the natural logarithm of the odds of each outcome.
}
\details{
\deqn{P(k | \mathbf{\pi}) = \frac{n!}{k! (n-k)!} \prod_{i=1}^k \pi_i}
}
\examples{
\donttest{
library(BI)
m=importBI(platform='cpu')
bi.dist.multinomial_logits(logits =  c(0.2, 0.3, 0.5), total_count = 10, sample = TRUE)
}
}
\seealso{
This is a wrapper of  \url{https://num.pyro.ai/en/stable/distributions.html#multinomiallogits}
}
