% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_sided_truncated_distribution.R
\name{bi.dist.two_sided_truncated_distribution}
\alias{bi.dist.two_sided_truncated_distribution}
\title{Two-Sided Truncated Distribution}
\usage{
bi.dist.two_sided_truncated_distribution(
  base_dist,
  low = 0,
  high = 1,
  validate_args = py_none(),
  name = "x",
  obs = py_none(),
  mask = py_none(),
  sample = FALSE,
  seed = 0,
  shape = c(),
  event = 0,
  create_obj = FALSE
)
}
\arguments{
\item{base_dist}{The base distribution to truncate.}

\item{low}{The lower bound for truncation.}

\item{high}{The upper bound for truncation.}

\item{sample}{Logical; if `TRUE`, returns JAX array of samples.  Defaults to `FALSE`.}

\item{create_obj}{Logical; if `TRUE`, returns the raw BI distribution object. Defaults to `FALSE`.}
}
\value{
- When \code{sample=FALSE}, a BI Two-Sided Truncated distribution object (for model building).

 - When \code{sample=TRUE}, a JAX array of samples drawn from the Two-Sided Truncated distribution (for direct sampling).

 - When \code{create_obj=TRUE}, the raw BI distribution object (for advanced use cases).
}
\description{
This distribution truncates a base distribution between two specified lower and upper bounds.

\deqn{f(x) = \begin{cases}
        \frac{p(x)}{P(\text{low} \le X \le \text{high})} & \text{if } \text{low} \le x \le \text{high} \\
        0 & \text{otherwise}
    \end{cases}}

where \deqn{p(x)} is the probability density function of the base distribution.
}
\examples{
\donttest{
library(BI)
m=importBI(platform='cpu')
bi.dist.two_sided_truncated_distribution(
base_dist = bi.dist.normal(0,1, create_obj = TRUE),
high = 0.5, low = 0.1, sample = TRUE)
}
}
\seealso{
\url{https://num.pyro.ai/en/stable/distributions.html#twosidedtruncateddistribution}
}
