% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load.R
\name{onLoad}
\alias{onLoad}
\title{Package Load}
\usage{
onLoad(libname = NULL, pkgname = "BI")
}
\arguments{
\item{libname}{Library name.}

\item{pkgname}{Package name.}
}
\value{
No return value, called for side effects.
}
\description{
Internal function to run when the package is loaded.
}
\keyword{internal}
