% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load.R
\name{onLoad}
\alias{onLoad}
\title{Package load hook}
\usage{
onLoad(libname = NULL, pkgname = "BI")
}
\arguments{
\item{libname}{Internal.}

\item{pkgname}{Internal.}
}
\description{
Internal function to run when the package is loaded.
}
\keyword{internal}
